Split the tail-call accumulator windows into their own file
